# src/main/resources/static/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "school--service",
    "version": "1.0.0"
  },
  "paths": {
...
    "/attendance": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateAttendance"
                }
              },
              "examples": {
                "request": {
                  "value": [
                    {
                      "student": 17,
                      "status": "late"
                    },
                    {
                      "student": 46,
                      "status": "late"
                    }
                  ]
                }
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "name": "token",
            "description": "auth token",
            "schema": {
              "type": "string"
            },
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/components/responses/AttendanceResponse"
          }
        },
        "operationId": "storeAttendance",
        "summary": "create an attendance",
        "description": "create an attendance"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Attendance": {
        "description": "",
        "required": [
          "date",
          "status"
        ],
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "description": "",
            "type": "string"
          },
          "status": {
            "description": "",
            "type": "string"
          }
        },
        "example": {
          "id": 67,
          "date": "2018-02-10T09:30Z"
        }
      },
      "CreateAttendance": {
        "description": "",
        "required": [
          "student",
          "status"
        ],
        "type": "object",
        "properties": {
          "student": {
            "description": "Student id",
            "type": "integer"
          },
          "status": {
            "description": "status",
            "enum": [
              "present",
              "missed",
              "late"
            ],
            "type": "string"
          }
        }
      }
    },
    "responses": {
...
      "AttendanceResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Attendance"
            },
            "examples": {
              "response": {
                "value": {
                  "id": 27,
                  "date": "2018-02-10T09:30Z"
                }
              }
            }
          }
        },
        "description": ""
      },
...
    }
  }
}